Redirect to the originally requested URL after logging in
This adds a redirect to the originally-requested URL once the user goes through the login process.

This is to restore this functionality to the Data Hub File Browser after the SSO logic was moved to this repository.

This uses the session to store the original URL as the authentication broker currently does not support equivalent functionality.